refactor singleton file name & add Final metaclass
