--- raspberrypi/common/singleton.py
@@ -1,16 +1,0 @@
-#!/usr/bin/python3
-# -*- coding: utf-8 -*-
-
-
-class Singleton(type):
-    """
-    Singleton Class
-    Usage:
-        class MyClass(metaclass=Singleton):
-            pass
-    """
-    _instances = {}
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        return cls._instances[cls]
